+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Base.Theme.FusMobilni"
         tools:targetApi="31">
         <activity
-            android:name=".activities.Register"
+            android:name=".activities.RegisterActivity"
             android:exported="false" />
         <activity
             android:name=".activities.LoginActivity"
