+++ app/src/main/res/drawable/toggle_button_selected.xml
@@ -1,0 +1,4 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+    <solid android:color="@color/primary_blue_900" />
+    <corners android:radius="16dp" />
+</shape>
