+++ app/src/main/res/drawable/toggle_button_selector.xml
@@ -1,0 +1,4 @@
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/toggle_button_selected" android:state_selected="true"/>
+    <item android:drawable="@drawable/toggle_button_unselected" android:state_selected="false"/>
+</selector>
