+++ app/src/main/res/drawable/toggle_button_unselected.xml
@@ -1,0 +1,4 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+    <solid android:color="@color/white" />
+    <corners android:radius="16dp" />
+</shape>
