+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,8 @@
     <string name="sign_in">Sign in</string>
     <string name="don_t_have_an_account">Don\'t have an account? <font color="#5669FF">Sign up</font></string>
     <string name="sign_up">Sign up\n</string>
+    <string name="name_placeholder">Milan Lazarevic\n</string>
+    <string name="repeat_password">Repeat password</string>
+    <string name="your_password">Your password</string>
+    <string name="have_an_account">Already have an account?<font color="#5669FF">Sign in</font></string>
 </resources>
